Microsoft_Purview_Purview_DG_Curator_Portal/requirements.txt
streamlit
pandas
requests
azure-identity
azure-purview-catalog
azure-purview-datamap
azure-core
python-dotenv


